+++ crates/guest-control-client/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [0.18.6](https://github.com/vm0-ai/vm0/compare/guest-control-client-v0.18.5...guest-control-client-v0.18.6) (2026-09-08)
+
 ## [0.18.5](https://github.com/vm0-ai/vm0/compare/guest-control-client-v0.18.4...guest-control-client-v0.18.5) (2026-09-07)